Screens in different folders. Consts refactor. Main entities creation
