+++ app/src/main/java/com/example/roomer/domain/model/entities/Room.kt
@@ -23,4 +23,5 @@ data class Room(
     val location: String = "Ordinary location",
     @SerializedName("title")
     val title: String = "Ordinary location",
+    val isLiked: Boolean = false,
 )
