+++ app/src/main/java/com/example/roomer/domain/usecase/signup/SignUpOneUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.roomer.domain.usecase.signup
 
 import com.example.roomer.data.repository.AuthRepositoryInterface
-import com.example.roomer.utils.ConstUseCase
+import com.example.roomer.utils.Constants
 import com.example.roomer.utils.Resource
 import java.io.IOException
 import kotlinx.coroutines.coroutineScope
@@ -34,7 +34,7 @@ class SignUpOneUseCase(
                 emit(Resource.Error.GeneralError(message = errMsg))
             }
         } catch (e: IOException) {
-            emit(Resource.Internet(ConstUseCase.internetErrorMessage))
+            emit(Resource.Internet(Constants.UseCase.internetErrorMessage))
         }
     }
 }
