+++ app/src/main/java/com/example/roomer/domain/usecase/signup/SignUpTwoUseCase.kt
@@ -2,7 +2,7 @@ package com.example.roomer.domain.usecase.signup
 
 import android.graphics.Bitmap
 import com.example.roomer.data.repository.AuthRepositoryInterface
-import com.example.roomer.utils.ConstUseCase
+import com.example.roomer.utils.Constants
 import com.example.roomer.utils.Resource
 import java.io.IOException
 import kotlinx.coroutines.coroutineScope
@@ -32,7 +32,7 @@ class SignUpTwoUseCase(
                 emit(Resource.Error.GeneralError(message = errMsg))
             }
         } catch (e: IOException) {
-            emit(Resource.Internet(ConstUseCase.internetErrorMessage))
+            emit(Resource.Internet(Constants.UseCase.internetErrorMessage))
         }
     }
 }
