+++ app/src/main/java/com/example/roomer/presentation/screens/search_screens/SearchRoomResultsViewModel.kt
@@ -1,9 +1,9 @@
-package com.example.roomer.presentation.screens
+package com.example.roomer.presentation.screens.search_screens
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.roomer.data.repository.AuthRepositoryInterface
-import com.example.roomer.domain.model.entities.User
+import com.example.roomer.domain.model.entities.Room
 import com.example.roomer.utils.LoadingStates
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
@@ -16,36 +16,33 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
 @HiltViewModel
-class SearchRoommateResultViewModel @Inject constructor(
+class SearchRoomResultsViewModel @Inject constructor(
     private val roomerRepository: AuthRepositoryInterface
 ) : ViewModel() {
-    private val _roommates = MutableStateFlow(emptyList<User>())
-    val roommates: StateFlow<List<User>> = _roommates
+    private val _rooms = MutableStateFlow(emptyList<Room>())
+    val rooms: StateFlow<List<Room>> = _rooms
     private val _loadingStates = MutableStateFlow(LoadingStates.Loading)
     val loadingState = _loadingStates.asStateFlow()
-    fun loadRoommates(
-        sex: String,
-        employment: String,
-        alcoholAttitude: String,
-        smokingAttitude: String,
-        sleepTime: String,
-        personalityType: String,
-        cleanHabits: String,
+
+    fun loadRooms(
+        monthPriceFrom: String,
+        monthPriceTo: String,
+        bedroomsCount: String,
+        bathroomsCount: String,
+        housingType: String,
     ) = effect {
         delay(2000)
         _loadingStates.value = LoadingStates.Loading
         coroutineScope {
-            val response = roomerRepository.getFilterRoommates(
-                sex,
-                employment,
-                alcoholAttitude,
-                smokingAttitude,
-                sleepTime,
-                personalityType,
-                cleanHabits
+            val response = roomerRepository.getFilterRooms(
+                monthPriceFrom,
+                monthPriceTo,
+                bedroomsCount,
+                bathroomsCount,
+                housingType,
             )
             if (response.isSuccessful) {
-                _roommates.value = response.body() ?: listOf()
+                _rooms.value = response.body() ?: listOf()
                 _loadingStates.value = LoadingStates.Success
             } else {
                 _loadingStates.value = LoadingStates.Error
