--- app/src/main/java/com/example/roomer/utils/ConstUseCase.kt
@@ -1,6 +1,0 @@
-package com.example.roomer.utils
-
-object ConstUseCase {
-    const val internetErrorMessage = "No internet connection!"
-    const val loginErrorName = "non_field_errors"
-}
